new e2e circle test
